Extract BuildOperationTreeQueries from BuildOperationFixture
in order for it to be usable without GradleExecuter
